+++ modules/core/builtin/cpp/Gateway.cpp
@@ -26,6 +26,7 @@
 #include "formatBuiltin.hpp"
 #include "execstrBuiltin.hpp"
 #include "evalBuiltin.hpp"
+#include "evalcBuiltin.hpp"
 #include "evalinBuiltin.hpp"
 #include "echoBuiltin.hpp"
 #include "versionBuiltin.hpp"
@@ -51,6 +52,7 @@ static const nlsGateway gateway[] =
     { "format", Nelson::CoreGateway::formatBuiltin, 1, 1 },
     { "execstr", Nelson::CoreGateway::execstrBuiltin, 1, 2 },
 	{ "eval", Nelson::CoreGateway::evalBuiltin, -1, 1 },
+    { "evalc", Nelson::CoreGateway::evalcBuiltin, -1, 1 },
     { "evalin", Nelson::CoreGateway::evalinBuiltin, -1, 2 },
     { "echo", Nelson::CoreGateway::echoBuiltin, 1, 1 },
     { "version", Nelson::CoreGateway::versionBuiltin, 2, 1 },