+++ modules/core/src/include/EvaluateCommand.hpp
@@ -27,6 +27,7 @@ namespace Nelson {
 	NLSCORE_IMPEXP bool EvaluateCommand(Evaluator *eval, std::wstring command, bool bCatch);
 	NLSCORE_IMPEXP bool EvaluateCommand(Evaluator *eval, std::string command, bool bCatch);
 	NLSCORE_IMPEXP ArrayOfVector EvaluateCommand(Evaluator *eval, int nLhs, std::wstring command, std::wstring catchCommand);
-	NLSCORE_IMPEXP ArrayOfVector EvaluateInCommand(Evaluator *eval, int nLhs, SCOPE_LEVEL scope,std::wstring command, std::wstring catchCommand);
+	NLSCORE_IMPEXP ArrayOfVector EvaluateInCommand(Evaluator *eval, int nLhs, SCOPE_LEVEL scope,std::wstring command);
+	NLSCORE_IMPEXP ArrayOfVector EvaluateConsoleCommand(Evaluator *eval, int nLhs, std::wstring command, std::wstring catchCommand);
 }
 //=============================================================================