+++ modules/interpreter/src/include/Evaluator.hpp
@@ -358,6 +358,8 @@ namespace Nelson {
          * Special case the single assignment statement 'A = B' for speed.
          */
         inline ArrayOf rhsExpressionSimple(ASTPtr t);
+		
+		void setInterface(Interface *_io);
         Interface* getInterface();
         /**
          * Process an AST to form an lvalue in an assignment statement.